refactor: check for `git` before installing/upgrade

+++ install-bproj.sh
@@ -4,6 +4,12 @@
 
 read -r -p "Are you sure you want to install/upgrade .bproj in $PWD? [y/N]" response
 if [[ "$response" =~ ^([yY][eE][sS]|[yY])$ ]]; then
+    GIT=`which git`
+    if [[ "$GIT" == '' ]]; then
+        echo "Missing \`git\` program. Please install \`git\` first."
+        echo "This command may help: \`apt-get install git\`"
+        exit
+    fi
     git clone https://github.com/Rp70/dotbproj.git dotbproj # Clone this repo
     rm -rdf ./.bproj # Remove the exising version
     mv dotbproj/.bproj ./ # Move the .bproj to your project folder
